refactor: Reorganize configuration components
This commit reorganizes the configuration components to improve the structure of the application. The configuration components have been moved to the `src/components/StudyMode/TemplateConfig` directory and renamed to follow a consistent naming convention. The import paths have been updated to reflect these changes.
