+++ src/components/StudyMode/TemplateConfig/ConfigureMCQMultipleBlock.js
@@ -1,202 +1,219 @@
 import React, { useState, useEffect } from 'react';
 import { useI18n } from '../../../i18n/I18nContext';
-import './SimpleTextConfig.css'; // Reusing for basic layout
-import './ConfigureMCQMultipleBlock.css'; // Specific styles
+import TransliterableText from '../../Common/TransliterableText';
+import Modal from '../../Common/Modal';
+import './ConfigureMCQMultipleBlock.css'; // Re-use common styles
 
-const ConfigureMCQMultipleBlock = ({ existingBlockData, onSave, onCancel, availableLanguages }) => {
-    const { t, language: currentUILanguage } = useI18n();
+const MCQMultipleBlockConfig = ({ block, onSave, onClose }) => {
+  const { t, currentLangKey } = useI18n();
+  const currentGlobalLang = currentLangKey || 'default';
 
-    const [blockTitle, setBlockTitle] = useState(existingBlockData?.title || {});
-    const [question, setQuestion] = useState(existingBlockData?.question || {});
-    // Options structure: [{ id: 'opt1', texts: {COSYenglish: 'Option 1'}, isCorrect: false }, ...]
-    const [options, setOptions] = useState(existingBlockData?.options || [{ id: `opt_${Date.now()}_0`, texts: {}, isCorrect: false }]);
+  const [title, setTitle] = useState(block.data.title || { default: '' });
+  const [question, setQuestion] = useState(block.data.question || { default: '' });
+  const [options, setOptions] = useState(block.data.options || []);
+  const [blockLang, setBlockLang] = useState(block.data.lang || '');
 
-    useEffect(() => {
-        if (existingBlockData) {
-            setBlockTitle(existingBlockData.title || {});
-            setQuestion(ensureAllLangs(existingBlockData.question || {}, availableLanguages, ''));
-            const loadedOptions = existingBlockData.options || [{ id: `opt_${Date.now()}_0`, texts: {}, isCorrect: false }];
-            setOptions(loadedOptions.map(opt => ({
-                ...opt,
-                texts: ensureAllLangs(opt.texts || {}, availableLanguages, '')
-            })));
-        } else {
-            setQuestion(ensureAllLangs({}, availableLanguages, ''));
-            setOptions([{ id: `opt_${Date.now()}_0`, texts: ensureAllLangs({}, availableLanguages, ''), isCorrect: false }]);
-        }
-    }, [existingBlockData, availableLanguages]);
-    
-    const ensureAllLangs = (textObj, langs, defaultValue = '') => {
-        const newTextObj = { ...textObj };
-        langs.forEach(lang => {
-            if (newTextObj[lang] === undefined || newTextObj[lang] === null) {
-                newTextObj[lang] = defaultValue;
-            }
-        });
-        return newTextObj;
+  useEffect(() => {
+    setTitle(block.data.title || { default: '' });
+    setQuestion(block.data.question || { default: '' });
+    // Ensure options have a feedback field, defaulting to an empty object if not present
+    setOptions((block.data.options || []).map(opt => ({ ...opt, feedback: opt.feedback || { default: '' } })));
+    setBlockLang(block.data.lang || '');
+  }, [block]);
+
+  const handleSave = () => {
+    const updatedBlockData = {
+      ...block.data,
+      title,
+      question,
+      options: options.map(opt => ({
+        ...opt,
+        // Ensure feedback is null if all its language fields are empty, or not present if object is empty
+        feedback: (opt.feedback && Object.values(opt.feedback).some(val => val && val.trim() !== '')) ? opt.feedback : null,
+      })),
+      lang: blockLang.trim() === '' ? null : blockLang.trim(),
     };
+    onSave({ ...block, data: updatedBlockData });
+    onClose();
+  };
 
-    const handleBlockTitleChange = (lang, value) => {
-        setBlockTitle(prev => ({ ...prev, [lang]: value }));
-    };
+  const handleMultilingualChange = (fieldSetter, currentFieldState, lang, value) => {
+    fieldSetter({ ...currentFieldState, [lang]: value });
+  };
 
-    const handleQuestionChange = (lang, value) => {
-        setQuestion(prev => ({ ...prev, [lang]: value }));
-    };
-
-    const handleOptionTextChange = (index, lang, value) => {
-        const newOptions = [...options];
-        if (!newOptions[index].texts) newOptions[index].texts = {};
-        newOptions[index].texts[lang] = value;
-        setOptions(newOptions);
-    };
-
-    const handleOptionCorrectChange = (index) => {
-        const newOptions = [...options];
-        newOptions[index].isCorrect = !newOptions[index].isCorrect;
-        setOptions(newOptions);
-    };
-
-    const addOption = () => {
-        // Limit number of options if desired, e.g., max 6
-        // if (options.length >= 6) {
-        //     alert(t('maxOptionsReached') || "Maximum of 6 options allowed.");
-        //     return;
-        // }
-        const newOption = { id: `opt_${Date.now()}_${options.length}`, texts: {}, isCorrect: false };
-        availableLanguages.forEach(lang => {
-            newOption.texts[lang] = '';
-        });
-        setOptions([...options, newOption]);
-    };
-
-    const removeOption = (index) => {
-        if (options.length > 2) { // Typically need at least 2 options for MCQ
-            setOptions(options.filter((_, i) => i !== index));
-        } else {
-            alert(t('minTwoOptionsMCQ') || 'At least two options are required for a multiple-choice question.');
-        }
-    };
-
-    const handleSaveConfiguration = () => {
-        const titleToSave = { ...blockTitle };
-        if (!titleToSave[currentUILanguage] && !titleToSave.COSYenglish) {
-            alert(t('alertTitleRequired') || `Please enter a title for the block.`);
-            return;
-        }
-        availableLanguages.forEach(lang => {
-            if (!titleToSave[lang]) {
-                titleToSave[lang] = titleToSave[currentUILanguage] || titleToSave.COSYenglish || 'Multiple Choice Question';
-            }
-        });
-        
-        const questionToSave = { ...question };
-        if (!Object.values(questionToSave).some(q => q.trim() !== '')) {
-            alert(t('alertMCQQuestionRequired') || 'Please enter the question text.');
-            return;
-        }
-         availableLanguages.forEach(lang => {
-            if (!questionToSave[lang]) {
-                questionToSave[lang] = questionToSave[currentUILanguage]?.trim() || questionToSave.COSYenglish?.trim() || '';
-            }
-        });
-
-        const validOptions = options.filter(opt => Object.values(opt.texts).some(text => text.trim() !== ''));
-        if (validOptions.length < 2) {
-            alert(t('alertMinTwoOptionsPopulatedMCQ') || 'Please provide text for at least two options.');
-            return;
-        }
-        if (!validOptions.some(opt => opt.isCorrect)) {
-            alert(t('alertMinOneCorrectOptionMCQ') || 'Please mark at least one option as correct.');
-            return;
-        }
-        
-        const processedOptions = validOptions.map(opt => {
-            const newTexts = { ...opt.texts };
-            availableLanguages.forEach(lang => {
-                if (!newTexts[lang] || newTexts[lang].trim() === '') {
-                     newTexts[lang] = newTexts[currentUILanguage]?.trim() || newTexts.COSYenglish?.trim() || `Option ${opt.id}`;
-                }
-            });
-            return { ...opt, texts: newTexts };
-        });
+  const handleOptionPropertyChange = (index, propertyName, value, lang = null) => {
+    const newOptions = [...options];
+    if (lang) { // For multilingual properties like 'texts' or 'feedback'
+      newOptions[index][propertyName] = {
+        ...(newOptions[index][propertyName] || {}),
+        [lang]: value
+      };
+    } else { // For direct properties like 'isCorrect'
+      newOptions[index][propertyName] = value;
+    }
+    setOptions(newOptions);
+  };
 
 
-        onSave({
-            id: existingBlockData?.id || `mcqm_${Date.now()}`,
-            type: 'comprehension/mcq-multiple',
-            title: titleToSave,
-            question: questionToSave,
-            options: processedOptions,
-        });
-    };
+  const addOption = () => {
+    setOptions([...options, {
+      id: `opt-${Date.now()}-${Math.random().toString(36).substr(2, 5)}`,
+      texts: { default: '' },
+      isCorrect: false,
+      feedback: { default: '' } // Initialize feedback field
+    }]);
+  };
 
-    return (
-        <div className="configure-mcq-multiple-block configure-simple-block">
-            <h4>{t('configureMCQMultipleTitle') || 'Configure Multiple Choice (Multiple Answers)'}</h4>
+  const removeOption = (index) => {
+    const newOptions = options.filter((_, i) => i !== index);
+    setOptions(newOptions);
+  };
 
-            <div className="form-group">
-                <label htmlFor="block-title-mcqm">{t('blockTitleLabel') || 'Block Title'} ({currentUILanguage}):</label>
-                <input
-                    type="text"
-                    id="block-title-mcqm"
-                    value={blockTitle[currentUILanguage] || blockTitle.COSYenglish || ''}
-                    onChange={(e) => handleBlockTitleChange(currentUILanguage, e.target.value)}
-                />
-            </div>
-
-            <div className="form-group">
-                <label htmlFor={`mcqm-question-${currentUILanguage}`}>{t('questionLabel', { langName: currentUILanguage.replace('COSY','') }) || `Question (${currentUILanguage.replace('COSY','')})`}:</label>
-                <textarea
-                    id={`mcqm-question-${currentUILanguage}`}
-                    value={question[currentUILanguage] || ''}
-                    onChange={(e) => handleQuestionChange(currentUILanguage, e.target.value)}
-                    rows="3"
-                />
-            </div>
-
-            <h5>{t('optionsLabel') || 'Options'}:</h5>
-            {options.map((option, index) => (
-                <div key={option.id || index} className="mcq-option-config-entry">
-                    <div className="form-group option-text-group">
-                        <label htmlFor={`mcqm-option-${index}-${currentUILanguage}`}>
-                            {t('optionTextLabel', { number: index + 1, langName: currentUILanguage.replace('COSY','') }) || `Option ${index + 1} Text (${currentUILanguage.replace('COSY','')})`}:
-                        </label>
-                        <input
-                            type="text"
-                            id={`mcqm-option-${index}-${currentUILanguage}`}
-                            value={option.texts?.[currentUILanguage] || ''}
-                            onChange={(e) => handleOptionTextChange(index, currentUILanguage, e.target.value)}
-                        />
-                    </div>
-                    <div className="form-group option-correct-group">
-                        <label htmlFor={`mcqm-correct-${index}`}>
-                            <input
-                                type="checkbox"
-                                id={`mcqm-correct-${index}`}
-                                checked={option.isCorrect}
-                                onChange={() => handleOptionCorrectChange(index)}
-                            />
-                            {t('correctAnswerCheckbox') || 'Correct Answer'}
-                        </label>
-                    </div>
-                    {options.length > 2 && (
-                        <button onClick={() => removeOption(index)} className="btn btn-danger btn-sm remove-option-btn">
-                            {t('removeOptionBtn') || 'Remove Option'}
-                        </button>
-                    )}
-                </div>
-            ))}
-            <button onClick={addOption} className="btn btn-secondary add-option-btn">
-                {t('addOptionBtn') || '+ Add Option'}
-            </button>
-
-            <div className="form-actions">
-                <button onClick={handleSaveConfiguration} className="btn btn-primary">{t('saveConfigurationBtn') || 'Save Configuration'}</button>
-                <button onClick={onCancel} className="btn btn-secondary">{t('cancelBtn') || 'Cancel'}</button>
-            </div>
+  // DEV_NOTE: Multilingual input is simplified to 'default' and current global language.
+  return (
+    <Modal
+      isOpen={true}
+      onClose={onClose}
+      title={t('configureBlockTitle', { blockName: block.typeName }) || `Configure ${block.typeName}`}
+    >
+      <div className="mcq-multiple-config-form block-config-form">
+        {/* Title */}
+        <div className="form-group">
+          <label className="unified-label"><TransliterableText text={t('config.titleDefaultLabel', 'Title (Default)')} />:</label>
+          <input
+            type="text"
+            value={title.default || ''}
+            onChange={(e) => handleMultilingualChange(setTitle, title, 'default', e.target.value)}
+          />
         </div>
-    );
+        {currentGlobalLang !== 'default' && (
+          <div className="form-group">
+            <label className="unified-label"><TransliterableText text={t('config.titleLangLabel', { lang: currentGlobalLang }, `Title (${currentGlobalLang})`)} />:</label>
+            <input
+              type="text"
+              value={title[currentGlobalLang] || ''}
+              onChange={(e) => handleMultilingualChange(setTitle, title, currentGlobalLang, e.target.value)}
+            />
+          </div>
+        )}
+
+        {/* Question */}
+        <div className="form-group">
+          <label className="unified-label"><TransliterableText text={t('config.questionDefaultLabel', 'Question (Default)')} />:</label>
+          <textarea
+            value={question.default || ''}
+            onChange={(e) => handleMultilingualChange(setQuestion, question, 'default', e.target.value)}
+            rows="3"
+          />
+        </div>
+        {currentGlobalLang !== 'default' && (
+          <div className="form-group">
+            <label className="unified-label"><TransliterableText text={t('config.questionLangLabel', { lang: currentGlobalLang }, `Question (${currentGlobalLang})`)} />:</label>
+            <textarea
+              value={question[currentGlobalLang] || ''}
+              onChange={(e) => handleMultilingualChange(setQuestion, question, currentGlobalLang, e.target.value)}
+              rows="3"
+            />
+          </div>
+        )}
+
+        {/* Options */}
+        <fieldset className="form-group options-fieldset">
+          <legend><TransliterableText text={t('config.mcqOptionsLabel', 'Options')} /></legend>
+          {options.map((option, index) => (
+            <div key={option.id || index} className="mcq-option-config-item">
+              <div className="option-inputs">
+                {/* Option Text */}
+                <div className="form-group option-text-group">
+                  <label htmlFor={`option-text-default-${index}`} className="unified-label"><TransliterableText text={t('config.optionTextDefaultLabel', 'Option Text (Default)')} />:</label>
+                  <input
+                    type="text"
+                    id={`option-text-default-${index}`}
+                    value={option.texts?.default || ''}
+                    onChange={(e) => handleOptionPropertyChange(index, 'texts', e.target.value, 'default')}
+                  />
+                </div>
+                {currentGlobalLang !== 'default' && (
+                  <div className="form-group option-text-group">
+                    <label htmlFor={`option-text-${currentGlobalLang}-${index}`} className="unified-label"><TransliterableText text={t('config.optionTextLangLabel', { lang: currentGlobalLang }, `Option Text (${currentGlobalLang})`)} />:</label>
+                    <input
+                      type="text"
+                      id={`option-text-${currentGlobalLang}-${index}`}
+                      value={option.texts?.[currentGlobalLang] || ''}
+                      onChange={(e) => handleOptionPropertyChange(index, 'texts', e.target.value, currentGlobalLang)}
+                    />
+                  </div>
+                )}
+
+                {/* Option Feedback */}
+                <div className="form-group option-feedback-group">
+                  <label htmlFor={`option-feedback-default-${index}`} className="unified-label"><TransliterableText text={t('config.optionFeedbackDefaultLabel', 'Feedback (Default)')} />:</label>
+                  <textarea
+                    id={`option-feedback-default-${index}`}
+                    value={option.feedback?.default || ''}
+                    onChange={(e) => handleOptionPropertyChange(index, 'feedback', e.target.value, 'default')}
+                    rows="2"
+                    placeholder={t('config.optionFeedbackPlaceholder', 'Optional feedback for this option')}
+                  />
+                </div>
+                {currentGlobalLang !== 'default' && (
+                  <div className="form-group option-feedback-group">
+                    <label htmlFor={`option-feedback-${currentGlobalLang}-${index}`} className="unified-label"><TransliterableText text={t('config.optionFeedbackLangLabel', { lang: currentGlobalLang }, `Feedback (${currentGlobalLang})`)} />:</label>
+                    <textarea
+                      id={`option-feedback-${currentGlobalLang}-${index}`}
+                      value={option.feedback?.[currentGlobalLang] || ''}
+                      onChange={(e) => handleOptionPropertyChange(index, 'feedback', e.target.value, currentGlobalLang)}
+                      rows="2"
+                      placeholder={t('config.optionFeedbackPlaceholder', 'Optional feedback for this option')}
+                    />
+                  </div>
+                )}
+
+                {/* Is Correct Checkbox */}
+                <div className="form-group option-correct-group">
+                  <label htmlFor={`option-correct-${index}`} className="checkbox-label unified-label"> {/* Also apply unified-label here, though it's a checkbox label */}
+                    <input
+                      type="checkbox"
+                      id={`option-correct-${index}`}
+                      checked={!!option.isCorrect} // Ensure it's a boolean
+                      onChange={(e) => handleOptionPropertyChange(index, 'isCorrect', e.target.checked)}
+                    />
+                    <TransliterableText text={t('config.isCorrectLabel', 'Correct Answer')} />
+                  </label>
+                </div>
+              </div>
+              <button type="button" onClick={() => removeOption(index)} className="btn btn-sm btn-danger remove-option-btn">
+                <TransliterableText text={t('removeOptionBtn', 'Remove Option')} />
+              </button>
+            </div>
+          ))}
+          <button type="button" onClick={addOption} className="btn btn-sm btn-secondary add-option-btn">
+            <TransliterableText text={t('addOptionBtn', '+ Add Option')} />
+          </button>
+        </fieldset>
+
+        {/* Block Language Override */}
+        <div className="form-group">
+          <label className="unified-label"><TransliterableText text={t('config.blockLangLabel', 'Language Override (e.g., COSYenglish)')} />:</label>
+          <input
+            type="text"
+            value={blockLang}
+            onChange={(e) => setBlockLang(e.target.value)}
+            placeholder={t('config.blockLangPlaceholder', 'Leave empty to use global language')}
+          />
+        </div>
+
+        <div className="modal-actions">
+          <button onClick={handleSave} className="btn btn-primary">
+            <TransliterableText text={t('saveConfigBtn', 'Save Configuration')} />
+          </button>
+          <button onClick={onClose} className="btn btn-secondary">
+            <TransliterableText text={t('cancelBtn', 'Cancel')} />
+          </button>
+        </div>
+      </div>
+    </Modal>
+  );
 };
 
-export default ConfigureMCQMultipleBlock;
+export default MCQMultipleBlockConfig;
