+++ src/components/StudyMode/TemplateConfig/ConfigureWritingBlock.js
